requirements.txt
tensorflow==2.4.1
numpy==1.19.5
scipy==1.4.1
tensorflow-hub==0.11.0
kafka-python==2.0.1
pyaudio==0.2.11
wavio==0.0.4